#438 cleanup/jsdoc
