+++ src/lib/how/sz-how-entity.component.ts
@@ -1,23 +1,20 @@
-import { Component, OnInit, Input, Inject, OnDestroy, Output, EventEmitter, ViewChild, HostBinding, ElementRef } from '@angular/core';
+import { Component, OnInit, Input, OnDestroy, Output, EventEmitter } from '@angular/core';
 import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { 
     EntityDataService as SzEntityDataService, 
-    SzEntityIdentifier, SzFeatureMode, SzHowEntityResponse, SzHowEntityResult, SzRecordId, SzRecordIdentifier, SzRecordIdentifiers, SzResolutionStep, SzResolvedEntity, SzVirtualEntity, SzVirtualEntityRecord, SzVirtualEntityResponse 
+    SzEntityIdentifier, SzFeatureMode, SzHowEntityResponse, SzHowEntityResult, SzRecordIdentifier, SzRecordIdentifiers, SzResolutionStep, SzResolvedEntity, SzVirtualEntity, SzVirtualEntityRecord, SzVirtualEntityResponse 
 } from '@senzing/rest-api-client-ng';
 import { SzConfigDataService } from '../services/sz-config-data.service';
 import { SzHowUIService } from '../services/sz-how-ui.service';
 import { 
-  SzHowFinalCardData, 
   SzVirtualEntityRecordsClickEvent, 
   SzResolvedVirtualEntity, 
-  SzResolutionStepGroup,
   SzResolutionStepDisplayType 
 } from '../models/data-how';
 import { Observable, Subject, take, takeUntil, zip, map } from 'rxjs';
 import { parseBool } from '../common/utils';
 import { v4 as uuidv4} from 'uuid';
 import { SzResolutionStepListItemType, SzResolutionStepNode } from '../models/data-how';
-
 
 /**
  * Display the "How" information for entity
@@ -37,21 +34,16 @@ import { SzResolutionStepListItemType, SzResolutionStepNode } from '../models/da
 export class SzHowEntityComponent implements OnInit, OnDestroy {
     /** subscription to notify subscribers to unbind */
     public unsubscribe$ = new Subject<void>();
-
-    //public finalCardsData: SzHowFinalCardData[];
     public finalCardsData: SzVirtualEntity[];
     private _data: SzHowEntityResult;
     private _virtualEntitiesById: Map<string, SzResolvedVirtualEntity>;
     private _featureTypesOrdered: string[] | undefined;
     private _resolutionSteps: Array<SzResolutionStep>;
     private _resolutionStepsByVirtualId: {[key: string]: SzResolutionStep};
-    //private _resolutionStepGroupsOld: Map<string, SzResolutionStepGroup> = new Map<string, SzResolutionStepGroup>();
-    //private _stepGroups: Map<string, SzResolutionStepGroup>     = new Map<string, SzResolutionStepGroup>();
     private _stepNodeGroups: Map<string, SzResolutionStepNode>  = new Map<string, SzResolutionStepNode>();
 
     private _isLoading                        = false;
     private _showNavigation                   = true;
-    private _hasChildStacksCached: boolean;
     private _entityId: SzEntityIdentifier;
     private _dataLoadedForId: SzEntityIdentifier;
     private _expandCardsWhenLessThan: number  = 2;
@@ -101,10 +93,6 @@ export class SzHowEntityComponent implements OnInit, OnDestroy {
     @Input() public set showNavigation(value: boolean | string) {
       this._showNavigation = parseBool(value);
     }
-
-    /*@HostBinding('class.has-child-stacks') get cssHasChildStacksClass(): boolean {
-      return this.hasChildStacks ? true : false;
-    }*/
 
     constructor(
         public entityDataService: SzEntityDataService,
@@ -218,45 +206,6 @@ export class SzHowEntityComponent implements OnInit, OnDestroy {
       }
     }
 
-    /*collapseAllSteps() {
-      this.howUIService.collapseAll();
-    }
-
-    public cooerceToStep(stepItem: SzResolutionStep | SzResolutionStepGroup): SzResolutionStep {
-      let retVal: SzResolutionStep | undefined;
-      if((stepItem as SzResolutionStep).stepNumber) {
-        retVal = (stepItem as SzResolutionStep);
-      }
-      return retVal;
-    }
-
-    public cooerceToStepGroup(stepItem: SzResolutionStep | SzResolutionStepGroup): SzResolutionStepGroup {
-      let retVal: SzResolutionStepGroup | undefined;
-      if((stepItem as SzResolutionStepGroup).resolutionSteps || (stepItem as SzResolutionStepGroup).interimSteps || (stepItem as SzResolutionStepGroup).mergeStep) {
-        retVal = (stepItem as SzResolutionStepGroup);
-      }
-      return retVal;
-    }
-
-    public getResolutionSteps(): Array<SzResolutionStep | SzResolutionStep[]> {
-      let retVal = [];
-      if(this._resolutionSteps && this._resolutionSteps.length > 1) {
-        retVal = this._resolutionSteps;
-      }
-      console.info('getResolutionSteps: ', retVal);
-      return retVal;
-    }
-
-    public getExpandedSteps() {
-      let retVal = this.howUIService.expandedNodes;
-      console.log(`getExpandedSteps(): `,retVal);
-    }
-    public getExpandedGroups() {
-      let retVal = this.howUIService.expandedGroups;
-      console.log(`getExpandedGroups(): `,retVal);
-    }*/
-
-
     public getStepNodesForInterimEntities(_rSteps?: Array<SzResolutionStep>): Map<string, SzResolutionStepNode> {
       let retVal        = new Map<string, SzResolutionStepNode>();
       let interimGroups = new Map<string, SzResolutionStepNode>();
@@ -358,73 +307,6 @@ export class SzHowEntityComponent implements OnInit, OnDestroy {
       console.log(`getParentContainingNode("${vId}") : `, retVal);
     }
 
-    /*
-    public getStepGroupsForInterimEntities(_rSteps?: Array<SzResolutionStep>): Map<string, SzResolutionStepGroup> {
-      let retVal        = new Map<string, SzResolutionStepGroup>();
-      let interimGroups = new Map<string, SzResolutionStepGroup>();
-
-      if(!_rSteps) {
-        _rSteps = this._resolutionSteps;
-      }
-      if(_rSteps && _rSteps.length > 1) {
-        // first initialize a group for each "merge" step
-        _rSteps.filter((resStep: SzResolutionStep) => {
-          return SzHowUIService.getResolutionStepCardType(resStep) === SzResolutionStepDisplayType.MERGE;
-        }).forEach((resStep: SzResolutionStep, stepArrIndex: number) => {
-          interimGroups.set(resStep.resolvedVirtualEntityId, {
-            id: resStep.resolvedVirtualEntityId,
-            mergeStep: resStep
-          });
-        });
-        // now go through all the steps
-        // and for each member in a interim entity 
-        // create a group
-        _rSteps.forEach((resStep: SzResolutionStep, stepArrIndex: number) => {
-          let vIdToLookFor  = resStep.resolvedVirtualEntityId;
-          interimGroups.forEach((value: SzResolutionStepGroup, key: string) => {
-            if(value.mergeStep && (
-              value.mergeStep.candidateVirtualEntity.virtualEntityId === vIdToLookFor || 
-              value.mergeStep.inboundVirtualEntity.virtualEntityId === vIdToLookFor)) {
-              
-              // step is a member of a merge
-              if(!retVal.has(vIdToLookFor)) {
-                // createa a group for it
-                retVal.set(vIdToLookFor, {
-                  id: vIdToLookFor,
-                  mergeStep: value.mergeStep,
-                  interimSteps: [resStep],
-                  virtualEntityIds: [resStep.candidateVirtualEntity.virtualEntityId, resStep.inboundVirtualEntity.virtualEntityId].filter((virtualEntityId: string) => {
-                    // make sure step is not another interim group
-                    return !interimGroups.has(virtualEntityId);
-                  })
-                });
-              }
-              // for each step who's resolvedId matches the
-              // iterim step id add it as a resolutionStep
-              if(this._resolutionStepsByVirtualId) {
-                // we only want to add steps if they are "Add Record" or "Create Entity" steps
-                if(this._resolutionStepsByVirtualId[resStep.candidateVirtualEntity.virtualEntityId]) {
-                  let stepType = SzHowUIService.getResolutionStepCardType(this._resolutionStepsByVirtualId[resStep.candidateVirtualEntity.virtualEntityId]);
-                  if(stepType === SzResolutionStepDisplayType.CREATE || stepType === SzResolutionStepDisplayType.ADD) { retVal.get(vIdToLookFor).interimSteps.push(this._resolutionStepsByVirtualId[resStep.candidateVirtualEntity.virtualEntityId]); }
-                }
-                if(this._resolutionStepsByVirtualId[resStep.inboundVirtualEntity.virtualEntityId]) {
-                  let stepType = SzHowUIService.getResolutionStepCardType(this._resolutionStepsByVirtualId[resStep.inboundVirtualEntity.virtualEntityId]);
-                  if(stepType === SzResolutionStepDisplayType.CREATE || stepType === SzResolutionStepDisplayType.ADD) { retVal.get(vIdToLookFor).interimSteps.push(this._resolutionStepsByVirtualId[resStep.inboundVirtualEntity.virtualEntityId]  ); }
-                }
-                if(!(retVal.get(vIdToLookFor).interimSteps && retVal.get(vIdToLookFor).interimSteps.length > 0)) {
-                  console.warn(`could not find resolution step for ${resStep.inboundVirtualEntity.virtualEntityId} or ${resStep.candidateVirtualEntity.virtualEntityId}`, this._resolutionStepsByVirtualId, retVal.get(vIdToLookFor));
-                }
-              } else {
-                console.warn('_resolutionStepsByVirtualId is null', this._resolutionStepsByVirtualId);
-              }
-            }
-          });
-        });
-      }
-      console.log('getStepGroupsForInterimEntities', retVal);
-      return retVal;
-    }*/
-
     private _nestInterimStepsForNodeGroup(stepNodeGroups: Map<string, SzResolutionStepNode>, stepNode: SzResolutionStepNode): SzResolutionStepNode {
       let _retVal: SzResolutionStepNode = Object.assign({}, stepNode);
       if(stepNode && stepNode.children && stepNode.itemType === SzResolutionStepListItemType.GROUP && stepNode.stepType === SzResolutionStepDisplayType.INTERIM) {
@@ -505,25 +387,6 @@ export class SzHowEntityComponent implements OnInit, OnDestroy {
       return _recursiveGroups;
     }
 
-    /*
-    private limitStepNodeGroupsToTopLevel(groups?: Map<string, SzResolutionStepNode>): Map<string, SzResolutionStepNode> {
-      let retVal = new Map();
-      for (let [key, value] of  groups.entries()) {
-        let isGroupMemberOfOtherGroup = this.stepIsMemberOfGroup(value.id, groups);
-        if(!isGroupMemberOfOtherGroup) {
-          retVal.set(key, value);
-        }
-      }
-      return retVal;
-    }*/
-
-    /*
-    public getStepNodeById(id) {
-      let _node = this.howUIService.getStepNodeById(id);
-      console.log(`getStepNodeById(${id})`, _node);
-      return _node;
-    }*/
-
     public get stepNodes(): Array<SzResolutionStepNode> {
       if(!this._resolutionSteps) { return undefined; }
       if(!this.howUIService.stepNodes || this.howUIService.stepNodes === undefined || (this.howUIService.stepNodes && this.howUIService.stepNodes.length <= 0)) {
@@ -569,49 +432,6 @@ export class SzHowEntityComponent implements OnInit, OnDestroy {
       }
       return this.howUIService.stepNodes;
     }
-    /*
-    public get fullyNestedList(): Array<SzResolutionStepNode> {
-      return this.getFullyNestedList();
-    }
-    public getFullyNestedList(): Array<SzResolutionStepNode> {
-      let _stepNodes = this.stepNodes;
-      let retVal: Array<SzResolutionStepNode> = [];
-      if(this.finalCardsData) {
-        // we have final card data
-        retVal = this.finalCardsData.map((fVirtualEntity: SzVirtualEntity) => {
-          let _resolvedVirtualEntity = this.virtualEntitiesById && this.virtualEntitiesById.has && this.virtualEntitiesById.has(fVirtualEntity.virtualEntityId) ? this.virtualEntitiesById.get(fVirtualEntity.virtualEntityId) : undefined;
-          let _fEntityAsStepNode: SzResolutionStepNode = Object.assign({
-            id: fVirtualEntity.virtualEntityId,
-            itemType: SzResolutionStepListItemType.FINAL,
-            stepType: SzResolutionStepDisplayType.FINAL,
-            isMemberOfGroup: false,
-            resolvedVirtualEntity: _resolvedVirtualEntity
-          }, fVirtualEntity);
-          // encode {dataSource: string, recordId: string, internalId: number} as "${dataSource}:${recordId}:${internalId}"
-          let _recsToCheckFor = _fEntityAsStepNode.records.map((ds)=> `${ds.dataSource}:${ds.recordId}:${ds.internalId}`);
-          if(_stepNodes && _stepNodes.length > 0 && _recsToCheckFor) {
-            // see if any step nodes should be children of this final card
-            let itemsAsChildren = _stepNodes.filter((stepNode: SzResolutionStepNode) => {
-              let _allRecordsForStepNode    = this.getRecordsForNode(true, stepNode);
-              // is all of the recordIds in the stepNode in the finalEntity
-              let _allRecordsInFinalEntity  = _allRecordsForStepNode.every((szVirtualRecord: SzVirtualEntityRecord) => {
-                return _recsToCheckFor.includes(`${szVirtualRecord.dataSource}:${szVirtualRecord.recordId}:${szVirtualRecord.internalId}`);
-              });
-              //stepNode = Object.assign(stepNode, {childRecords: _allRecordsForStepNode});
-              return _allRecordsInFinalEntity;
-            });
-            _fEntityAsStepNode.children = itemsAsChildren;
-          }
-          return _fEntityAsStepNode;
-        });
-      }
-      return retVal;
-      //console.log(`getFullyNestedList(): `, retVal, _stepNodes);
-    }
-    public getFullyNestedListDebug() {
-      let retVal = this.getFullyNestedList();
-      console.log(`getFullyNestedListDebug(): `, retVal);
-    }*/
 
     public getDefaultStepNodeGroups(_rSteps?: Array<SzResolutionStep>): Map<string, SzResolutionStepNode> {
       let retVal = new Map<string, SzResolutionStepNode>();
@@ -684,78 +504,13 @@ export class SzHowEntityComponent implements OnInit, OnDestroy {
       return retVal;
     }
 
-    /*
-    public getDefaultStepGroups(_rSteps?: Array<SzResolutionStep>): Map<string, SzResolutionStepGroup> {
-      let retVal = new Map<string, SzResolutionStepGroup>();
-      if(!_rSteps) {
-        _rSteps = this._resolutionSteps;
-      }
-      if(_rSteps && _rSteps.length > 1) {
-        let _stepGroupsForInterimEntities = this.getStepGroupsForInterimEntities(_rSteps);
-        let _currentGroupId   = uuidv4();
-
-        let _filteredSteps = _rSteps.filter((resStep: SzResolutionStep, stepArrIndex: number) => {
-          // if step is a member of an interim entity we should 
-          // exclude it from list
-          return !this.isStepMemberOfGroup(resStep, _stepGroupsForInterimEntities) && !(SzHowUIService.getResolutionStepCardType(resStep) === SzResolutionStepDisplayType.MERGE);
-        });
-        _filteredSteps.forEach((resStep: SzResolutionStep, stepArrIndex: number) => {
-          // check to see if we have any "top-level" collapsible groups
-          let futureStepType  = _rSteps[(stepArrIndex + 1)]  ? SzHowUIService.getResolutionStepCardType((_rSteps[(stepArrIndex + 1)] as SzResolutionStep), resStep.stepNumber) : undefined;
-          let currentStepType = _rSteps[stepArrIndex]        ? SzHowUIService.getResolutionStepCardType((_rSteps[ stepArrIndex     ] as SzResolutionStep), resStep.stepNumber) : undefined;
-          let lastStepType    = _rSteps[(stepArrIndex - 1)]  ? SzHowUIService.getResolutionStepCardType((_rSteps[(stepArrIndex - 1)] as SzResolutionStep), resStep.stepNumber) : undefined;
-
-          if(currentStepType !== SzResolutionStepDisplayType.ADD){ 
-            //_resolutionStepsWithGroups.push(resStep);
-          } else {
-            // current type is "add"
-              // check if previous step was add
-              // if so add this item to previous item
-              if(lastStepType === SzResolutionStepDisplayType.ADD) {
-                retVal.get(_currentGroupId).resolutionSteps.push(resStep);
-                this.howUIService.collapseStep(resStep.resolvedVirtualEntityId);
-                //console.log(`${stepArrIndex} | previous step was add operation. append`, retVal.get(_currentGroupId));
-                //(_resolutionStepsWithGroups[(_resolutionStepsWithGroups.length - 1)] as SzResolutionStep[]).push(resStep);
-              } else if(futureStepType === SzResolutionStepDisplayType.ADD) {
-                // this is the first "add record" in a sequence of at least
-                // two add records
-                _currentGroupId   = uuidv4();
-                retVal.set(_currentGroupId, {
-                  id: _currentGroupId,
-                  isStackGroup: true,
-                  resolutionSteps: []
-                });
-
-                retVal.get(_currentGroupId).arrayIndex = stepArrIndex;
-                retVal.get(_currentGroupId).resolutionSteps.push(resStep);
-                this.howUIService.collapseStep(resStep.resolvedVirtualEntityId);
-                //_resolutionStepsWithGroups.push([resStep]);
-                //console.log(`${stepArrIndex} | first add operation in series`, retVal.get(_currentGroupId));
-              } else {
-                // this is a "singular" "add record" step, do not group it
-                //_resolutionStepsWithGroups.push(resStep);
-              }
-              retVal.get(_currentGroupId).virtualEntityIds = retVal.get(_currentGroupId).resolutionSteps.map((rStep: SzResolutionStep) => { return rStep.resolvedVirtualEntityId; });
-          }
-        });
-
-        _stepGroupsForInterimEntities.forEach((group, key) => {
-          retVal.set(key, group);
-          this.howUIService.collapseStep(group.id);
-        });
-
-      }
-      console.info('getDefaultStepGroups: ', retVal);
-      return retVal;
-    }*/
-
-    private getGroupForMemberStep(step: SzResolutionStep, groups?: Map<string, SzResolutionStepGroup>): SzResolutionStepNode {
-      let _retVal: SzResolutionStepGroup;
+    private getGroupForMemberStep(step: SzResolutionStep, groups?: Map<string, SzResolutionStepNode>): SzResolutionStepNode {
+      let _retVal: SzResolutionStepNode;
       if(!groups) { groups = this._stepNodeGroups; }
       if(groups && step) {
         let _idToLookFor = step.resolvedVirtualEntityId;
         let _sk = false;
-        groups.forEach((groupToSearch: SzResolutionStepGroup, key: string) => {
+        groups.forEach((groupToSearch: SzResolutionStepNode, key: string) => {
           if(!_sk && groupToSearch.virtualEntityIds && groupToSearch.virtualEntityIds.indexOf(_idToLookFor) > -1 || groupToSearch.id === _idToLookFor) {
             _retVal = groupToSearch;
           }
@@ -763,24 +518,6 @@ export class SzHowEntityComponent implements OnInit, OnDestroy {
       }
       return _retVal;
     }
-
-    /*
-    public getNestedStepGroup(vId) {
-      let _stepGroups = this.getDefaultStepNodeGroups(this._resolutionSteps);
-      let _retVal;
-      if(_stepGroups && _stepGroups.has(vId)) {
-        _retVal = this._nestInterimStepsForNodeGroup(_stepGroups, _stepGroups.get(vId));
-        _retVal.virtualEntityIds = this.getVirtualEntityIdsForNode(_retVal);
-        console.log(`getNestedStepGroup(${vId})`, _retVal, _stepGroups);
-        return _retVal;
-      }
-      return undefined;
-    }*/
-
-    /*
-    public getResolutionStepsAsNodesDebug() {
-      return this.getResolutionStepsAsNodes(this._resolutionSteps, this._stepNodeGroups);
-    }*/
 
     public getResolutionStepsAsNodes(steps: SzResolutionStep[], groups: Map<string, SzResolutionStepNode>): SzResolutionStepNode[] {
       let retVal: SzResolutionStepNode[] = [];
@@ -852,193 +589,12 @@ export class SzHowEntityComponent implements OnInit, OnDestroy {
       return retVal;
     }
 
-    /*private getStepNodeForChild(step: SzResolutionStep, groups?: Map<string, SzResolutionStepNode>): SzResolutionStepGroup {
-      let _retVal: SzResolutionStepNode;
-      if(!groups) { groups = this._stepNodeGroups; }
-      if(groups && step) {
-        let _idToLookFor = step.resolvedVirtualEntityId;
-        let _sk = false;
-        groups.forEach((groupToSearch: SzResolutionStepNode, key: string) => {
-          if(!_sk && groupToSearch.virtualEntityIds && groupToSearch.virtualEntityIds.indexOf(_idToLookFor) > -1 || groupToSearch.id === _idToLookFor) {
-            _retVal = groupToSearch;
-          }
-        });
-      }
-      return _retVal;
-    }*/
-
-    /*
-    private isStepChildOfStep(step: SzResolutionStep, groups?: Map<string, SzResolutionStepNode>) {
-      if(this.getStepNodeForChild(step, groups) !== undefined) {
-        return true;
-      }
-      return false;
-    }*/
-
-    private isStepMemberOfGroup(step: SzResolutionStep, groups?: Map<string, SzResolutionStepGroup>) {
+    private isStepMemberOfGroup(step: SzResolutionStep, groups?: Map<string, SzResolutionStepNode>) {
       if(this.getGroupForMemberStep(step, groups) !== undefined) {
         return true;
       }
       return false;
     }
-
-    /*
-    public getDefaultResolutionStepGroups(_rSteps?: Array<SzResolutionStep>): Map<string, SzResolutionStepGroup> {
-      let retVal = new Map<string, SzResolutionStepGroup>();
-      if(!_rSteps) {
-        _rSteps = this._resolutionSteps;
-      }
-      // create "groups" for multiple sequential "add record" steps
-      if(_rSteps && _rSteps.length > 1) {
-        //let _resolutionStepsWithGroups: Array<SzResolutionStep | SzResolutionStep[]> = [];
-        let _currentGroupId   = uuidv4();
-
-        _rSteps.forEach((resStep: SzResolutionStep, stepArrIndex: number) => {
-          let futureStepType  = _rSteps[(stepArrIndex + 1)]  ? SzHowUIService.getResolutionStepCardType((_rSteps[(stepArrIndex + 1)] as SzResolutionStep), resStep.stepNumber) : undefined;
-          let currentStepType = _rSteps[stepArrIndex]        ? SzHowUIService.getResolutionStepCardType((_rSteps[ stepArrIndex     ] as SzResolutionStep), resStep.stepNumber) : undefined;
-          let lastStepType    = _rSteps[(stepArrIndex - 1)]  ? SzHowUIService.getResolutionStepCardType((_rSteps[(stepArrIndex - 1)] as SzResolutionStep), resStep.stepNumber) : undefined;
-
-          if(currentStepType !== SzResolutionStepDisplayType.ADD){ 
-            //_resolutionStepsWithGroups.push(resStep);
-          } else {
-            // current type is "add"
-              // check if previous step was add
-              // if so add this item to previous item
-              if(lastStepType === SzResolutionStepDisplayType.ADD) {
-                retVal.get(_currentGroupId).resolutionSteps.push(resStep);
-                //this.howUIService.collapseStep(resStep.resolvedVirtualEntityId);
-                //console.log(`${stepArrIndex} | previous step was add operation. append`, retVal.get(_currentGroupId));
-                //(_resolutionStepsWithGroups[(_resolutionStepsWithGroups.length - 1)] as SzResolutionStep[]).push(resStep);
-              } else if(futureStepType === SzResolutionStepDisplayType.ADD) {
-                // this is the first "add record" in a sequence of at least
-                // two add records
-                _currentGroupId   = uuidv4();
-                retVal.set(_currentGroupId, {
-                  id: _currentGroupId,
-                  isStackGroup: true,
-                  resolutionSteps: []
-                });
-
-                retVal.get(_currentGroupId).arrayIndex = stepArrIndex;
-                retVal.get(_currentGroupId).resolutionSteps.push(resStep);
-                //this.howUIService.collapseStep(resStep.resolvedVirtualEntityId);
-                //_resolutionStepsWithGroups.push([resStep]);
-                //console.log(`${stepArrIndex} | first add operation in series`, retVal.get(_currentGroupId));
-              } else {
-                // this is a "singular" "add record" step, do not group it
-                //_resolutionStepsWithGroups.push(resStep);
-              }
-              retVal.get(_currentGroupId).virtualEntityIds = retVal.get(_currentGroupId).resolutionSteps.map((rStep: SzResolutionStep) => { return rStep.resolvedVirtualEntityId; });
-          }
-        });
-        //retVal  = _resolutionStepsWithGroups;
-      }
-      //console.info('getResolutionStepGroups: ', retVal);
-      return retVal;
-    }*/
-
-    /*
-    public toggleCollapsedState(step: SzResolutionStep | SzResolutionStepGroup) {
-      let vId = (step as SzResolutionStepGroup).id ? (step as SzResolutionStepGroup).id : (step as SzResolutionStep).resolvedVirtualEntityId ? (step as SzResolutionStep).resolvedVirtualEntityId : undefined;
-      if(vId) {
-        if(this.howUIService.isExpanded(vId)){
-          this.howUIService.collapse(vId);
-        } else {
-          this.howUIService.expand(vId);
-        }
-      }
-    }
-
-    public toggleGroupCollapsedState(gId: SzResolutionStepGroup) {
-      this.howUIService.toggleExpansion(undefined, gId.id);
-    }*/
-
-    /*
-    public isParentEntityHidden(member: SzResolutionStep | SzResolutionStepGroup) {
-      let finalEntity = this.getFinalEntityFromMember(member);
-      let retVal = false;
-      if(finalEntity) {
-        retVal = !this.howUIService.isFinalEntityExpanded(finalEntity.virtualEntityId)
-      }
-      //console.log(`isParentEntityHidden(${retVal})`, finalEntity, member);
-      return retVal;
-    }
-
-    private getFinalEntityFromMember(member: SzResolutionStep | SzResolutionStepGroup) {
-      let isGroup = (member as SzResolutionStepGroup).virtualEntityIds ? true : false;
-      // you can dance if you want to !!! ... umm.. umm.. bumm.. SAFETY DANCE!!  OOOOOOhhhOOOhhhOooo SAFETY DANCE!
-      if(!(this.finalCardsData && this.finalCardsData.find)) {
-        return undefined;
-      }
-      if(isGroup) {
-        // check to see if all virtual entity ids are in a particular group
-        let memberAsGroup = (member as SzResolutionStepGroup);
-        let parentFinal = this.finalCardsData.find((fEnt: SzVirtualEntity) => {
-          let idToLookFor = fEnt.virtualEntityId;
-          return memberAsGroup.virtualEntityIds && memberAsGroup.virtualEntityIds.includes(idToLookFor);
-        });
-        return parentFinal;
-      } else {
-        // is step
-        let memberAsStep = (member as SzResolutionStep);
-        let parentFinal = this.finalCardsData.find((fEnt: SzVirtualEntity) => {
-          // look through fEnt.records for record matches??!?! (yeah this is dirty)
-          // but no better way to tie these two together for now
-          let _recordsToLookFor = memberAsStep.candidateVirtualEntity.records.concat(memberAsStep.inboundVirtualEntity.records);
-          let _hasAllRecords    = _recordsToLookFor.every((record) => {
-            // convert records to strings for comparison
-            let _recKeys  = fEnt.records.map((r)=>{ return r.dataSource+':'+r.recordId+':'+r.internalId});
-            let hasRecord = _recKeys.includes(record.dataSource+':'+record.recordId+':'+record.internalId);
-            return hasRecord;
-          });
-          return _hasAllRecords;
-        });
-        //console.log(`\tis step(${memberAsStep.resolvedVirtualEntityId})`, 
-        //parentFinal, memberAsStep, 
-        //this.finalCardsData, this._data.finalStates);
-
-        return parentFinal;
-      }
-      return undefined;
-    }
-
-    public getFinalEntityFromMemberByStepNumber(stepNumber: number){
-      // first get step
-      let stepsOrGroups = this.stepsList;
-      let member = stepsOrGroups.find((stepOrGroup) => {
-        let isGroup       = stepOrGroup.itemType === SzResolutionStepListItemType.GROUP ? true : false;
-
-        if(isGroup) {
-          return stepOrGroup.children.find((step) => {
-            return step.stepNumber === stepNumber;
-          })
-        } else {
-          return stepOrGroup.stepNumber === stepNumber;
-        }
-        return undefined;
-      });
-      if(member) {
-        // now get parent group
-        let finalEntity = this.getFinalEntityFromMember(member);
-        console.log('getFinalEntityFromMemberByStepNumber: found step in stepsList: ', member, finalEntity);
-      }
-    }*/
-
-    /*
-    public isGroupExpanded(grp: SzResolutionStepNode) {
-      let vId = grp.id ? grp.id : grp.resolvedVirtualEntityId ? grp.resolvedVirtualEntityId : undefined;
-      if(vId) {
-        return this.howUIService.isGroupExpanded(vId);
-      }
-      return false;
-    }
-
-    public getListItemType(item: SzResolutionStep | SzResolutionStepGroup): SzResolutionStepListItemType {
-      let _retVal = SzHowUIService.getResolutionStepListItemType(item);
-      let stepNumber = (item as SzResolutionStep).stepNumber ? (item as SzResolutionStep).stepNumber : (item as SzResolutionStepGroup).id
-      //console.log(`${stepNumber} ${_retVal}`, item);
-      return _retVal;
-    }*/
 
     public stepIsMemberOfGroup(virtualEntityId: string, groups: Map<string, SzResolutionStepNode>) {
       let _retVal   = false;
@@ -1064,13 +620,6 @@ export class SzHowEntityComponent implements OnInit, OnDestroy {
       return _g ? _g : undefined;
     }
 
-    /*public getResolutionStepGroupIdByMemberVirtualIdDebug(virtualEntityId: string) {
-      let groups  = this.getStepNodeGroupsRecursively(this._resolutionSteps, this._stepNodeGroups);
-      let _retVal = this.getResolutionStepGroupIdByMemberVirtualId(virtualEntityId, groups);
-      console.info(`getResolutionStepGroupIdByMemberVirtualIdDebug('${virtualEntityId}'): `,_retVal, groups);
-      return _retVal;
-    }*/
-
     private getResolutionStepGroupIdByMemberVirtualId(virtualEntityId: string, groups: Map<string, SzResolutionStepNode>) {
       let retVal;
       if(groups) {
@@ -1093,73 +642,6 @@ export class SzHowEntityComponent implements OnInit, OnDestroy {
       return undefined;
     }
 
-    /*
-    public get resolutionStepsWithGroupsOld(): Array<SzResolutionStep | SzResolutionStepNode> {
-      let retVal = this.getResolutionStepsWithGroups(this._resolutionSteps, this._resolutionStepGroupsOld);
-      return retVal;
-    }*/
-
-    /*
-    public get stepsList(): Array<SzResolutionStepNode> {
-      if(!this.howUIService.stepNodes || this.howUIService.stepNodes === undefined) {
-        this.howUIService.stepNodes   = this.getResolutionStepsWithGroups(this._resolutionSteps, this._stepNodeGroups);
-      }
-      return this.howUIService.stepNodes;
-    }
-
-    public getResolutionStepsWithGroups(steps: SzResolutionStep[], groups: Map<string, SzResolutionStepNode>): Array<SzResolutionStepNode> {
-      let retVal = [];
-      // create "groups" for multiple sequential "add record" steps
-      if(steps && steps.length > 0) {
-        let _resolutionStepsWithGroups: Array<SzResolutionStepNode> = [];
-
-        steps.forEach((resStep: SzResolutionStep, stepArrIndex: number) => {
-          // check to see if step is member of group
-          let _stepIsMemberOfGroup        = this.stepIsMemberOfGroup(resStep.resolvedVirtualEntityId, groups);
-          let _stepHasMembers             = this.stepHasMembers(resStep.resolvedVirtualEntityId, groups);
-
-          // check whether or not step is member of group
-          if(!_stepIsMemberOfGroup && !_stepHasMembers) {
-            // item is not member of group or interim step group
-            // add item to array
-            _resolutionStepsWithGroups.push(Object.assign({
-              id: resStep.resolvedVirtualEntityId,
-              itemType: SzResolutionStepListItemType.STEP,
-              displayType: this.getListItemType(resStep)
-            }, resStep));
-          } else {
-            // item is in group, if group has not been added already add it
-            let _stepGroup            = _stepIsMemberOfGroup ? this.getResolutionStepGroupByMemberVirtualId(resStep.resolvedVirtualEntityId, groups) : this.getResolutionStepGroupById(resStep.resolvedVirtualEntityId, groups);
-            let groupAlreadyInArray   = _resolutionStepsWithGroups.includes(_stepGroup);
-            if(!groupAlreadyInArray) { 
-              _resolutionStepsWithGroups.push( _stepGroup ); 
-            }
-          }
-        });
-        retVal  = _resolutionStepsWithGroups;
-      }
-      //console.info('resolutionStepsWithGroups: ', retVal);
-      return retVal;
-    }*/
-
-    /*
-    getStepNumberByVirtualEntityId(virtualEntityId: string) {
-        let retVal = undefined;
-        if(this._resolutionStepsByVirtualId && this._resolutionStepsByVirtualId[virtualEntityId]) {
-            retVal = this._resolutionStepsByVirtualId[virtualEntityId].stepNumber;
-        }
-        return retVal;
-    }
-
-    public getConstructionStepForVirtualEntity(virtualEntityId: string) {
-        let retVal = undefined;
-        if(this._resolutionStepsByVirtualId && this._resolutionStepsByVirtualId[virtualEntityId]) {
-            // there is a step entry for the virtualEntityId passed to this function
-            retVal = this._resolutionStepsByVirtualId[virtualEntityId];
-        }
-        return retVal;
-    }*/
-
     getData(entityId: SzEntityIdentifier): Observable<SzHowEntityResponse> {
         return this.entityDataService.howEntityByEntityID(
             this.entityId as number
@@ -1173,21 +655,6 @@ export class SzHowEntityComponent implements OnInit, OnDestroy {
         });
     }
 
-    /*
-    public get hasChildStacks(): boolean {
-      let retVal = false;
-      let _d = this.stepNodes;
-      if(_d && _d.length > 0) {
-        if(this._hasChildStacksCached === undefined) {
-            this._hasChildStacksCached = _d.some((childItem) => {
-                return (childItem as SzResolutionStepNode).itemType === SzResolutionStepListItemType.STACK;
-            })
-        }
-        return this._hasChildStacksCached ? true : false;
-      }
-      return retVal;
-    }*/
-
     /**
      * unsubscribe when component is destroyed
      */
@@ -1195,15 +662,6 @@ export class SzHowEntityComponent implements OnInit, OnDestroy {
         this.unsubscribe$.next();
         this.unsubscribe$.complete();
     }
-
-    /*
-    public isResolutionStep(value: SzResolutionStep | SzResolutionStepGroup): boolean {
-      let retVal = false;
-      if((value as SzResolutionStep).stepNumber !== undefined) {
-        retVal = true;
-      }
-      return retVal;
-    }*/
 
     /** 
      * this is the method that does the heavy lifting for getting ALL the data
